Update print a float.c

+++ C vs Assembly/print a float.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(){
-int number=2.5;
+float number=2.5;
   printf("%f",number);
   return 0;
 }
